+++ Modules/Config/ConfigModule.lua
@@ -409,7 +409,6 @@ local options = {
                 profile.notifyForRareUpgrade = not profile.notifyForRareUpgrade
             end
         },
---@retail@
         objectiveTrackerHeader = {
             order = 31,
             name = L["Header - Objective Tracker"],
@@ -462,7 +461,6 @@ local options = {
                 ObjectiveTrackerModule:OnEvent("CONFIG_CHANGED")
             end
         },
---@end-retail@
         dataVersionHeader = {
             order = 35,
             type = "header",