+++ Changelog.md
@@ -1,5 +1,6 @@
 # 12.0.1-20260323-2
 * Fix bug in MapModule in Classic.
+* Implement objective tracker in Classic.
 
 # 12.0.1-20260323-1
 * Declutter map, to avoid too close pins.